requirements.txt
flask==2.1.1
werkzeug==2.0.3
requests==2.26.0
openai==0.27.0
streamlit==1.8.0
langchain==0.0.158
faiss-cpu==1.7.3
python-dotenv==0.20.0